Swap Destroy Test To Do: User Test

+++ app/Http/Controllers/SwapController.php
@@ -130,7 +130,7 @@ class SwapController extends Controller
     public function destroy(Swap $swap)
     {
         // Delete the image
-        if ($swap->image !== 'empty.svg') {
+        if ($swap->image !== 'empty.svg' && file_exists(public_path('images/food/' . $swap->image))) {
             unlink(public_path('images/food/' . $swap->image));
         }
